Implement better handling of ◊txt elements
Previously, if a ◊txt tag contained a ◊txt-noescape tag, you would get
an error since the latter are excluded in the first pass of decoding at
the ◊root level and would thus not exist in pure string form for the
containing ◊txt tag’s `string-append` call. The new method simply
passes all contained elements through untouched.

When targeting LaTeX/PDF, it is necessary that the whole doc end up as
a single txexpr containing only strings. This change delegates
responsibility for this to the ◊root function.